+++ CHANGELOG.md
@@ -1,7 +1,33 @@
-## 0.1.0
+
+## v0.3.1
+2024-01-15
+### Perf
+
+* 添加 divisionForUpload 和 sliceSizeForUpload 参数到 ObjectStoragePutObjectRequest
+
+
+## v0.3.0
+2024-01-15
+### Chore
+
+* Change dep cross_file to platform
+
+
+## v0.2.0
+2024-01-11
+### Chore
+
+* 代码格式化
+* 添加 CHANGELOG.md 生成脚本
+
+### Refactor
+
+* 使用 XFile file 替换 String filePath 以兼容多平台
+
+
+## v0.1.0
+2022-11-29
+### Refactor
 
 * 调整回调接口参数
 
-## 0.0.1
-
-* Initial release.